# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Envoitures Rest Api",
        "version": "0.1"
    },
    "paths": {
...
        "/api/bookings": {
            "get": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Get List of bookings, for filters use: <<attributeName>>[<<operator>>]=<<value>>, operatorList=[eq,lt,lte,gt,gte] e.g userId[eq]=3,\nAnd to include other properties like user or ride in the response, use : include<<propertyName>>=true, e.g includeRide=true, includeUser=true",
                "operationId": "eb7c332e8ba161904a1c0071040f61a5",
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rows": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "userId": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "rideId": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "suggestedPrice": {
                                                        "type": "number",
                                                        "example": "32"
                                                    },
                                                    "validatedAt": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2001-12-11"
                                                    },
                                                    "payment": {
                                                        "type": "string",
                                                        "example": "pending"
                                                    },
                                                    "paidAt": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2001-12-11"
                                                    },
                                                    "fee": {
                                                        "type": "number",
                                                        "example": "10"
                                                    },
                                                    "isValidated": {
                                                        "type": "number",
                                                        "example": "0"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2021-12-11T09:25:53.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Bookings"
                ],
                "summary": "Create Booking",
                "operationId": "e33fa5a6507581f9b1e0d710a9a130d1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "userId": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "rideId": {
                                                "type": "number",
                                                "example": "1"
                                            },
                                            "price": {
                                                "type": "nullable",
                                                "example": "32"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "userId": "1",
                                    "rideId": "1",
                                    "price": "32"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "rideId": {
                                            "type": "number",
                                            "example": "1"
                                        },
                                        "suggestedPrice": {
                                            "type": "number",
                                            "example": "32"
                                        },
                                        "price": {
                                            "type": "nullable",
                                            "example": "32"
                                        },
                                        "validatedAt": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2001-12-11"
                                        },
                                        "payment": {
                                            "type": "string",
                                            "example": "pending"
                                        },
                                        "paidAt": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "2001-12-11"
                                        },
                                        "fee": {
                                            "type": "number",
                                            "example": "10"
                                        },
                                        "isValidated": {
                                            "type": "number",
                                            "example": "0"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}